Cast size_t to DWORD on Windows

+++ common/randombytes.c
@@ -90,7 +90,7 @@ static int randombytes_win32_randombytes(void *buf, const size_t n) {
         return -1;
     }
 
-    tmp = CryptGenRandom(ctx, n, (BYTE *)buf);
+    tmp = CryptGenRandom(ctx, (DWORD)n, (BYTE *)buf);
     if (tmp == FALSE) {
         return -1;
     }
